requirements.txt
fastapi
uvicorn
python-dotenv
openai
pytesseract
pdf2image
Pillow
PyMuPDF